refactor(frontend): Implement new API client for service interactions
Transition all service functions in `services.ts` to use the new `apiGet`, `apiPost`, `apiPut`, and `apiDelete` helper functions. This centralizes API request logic, enhancing maintainability and consistency across the frontend.
Relocate `services.ts` and `SNSPage.tsx` to the `packages/frontend` directory as part of the monorepo migration.
