+++ .env.example
@@ -1,5 +1,6 @@
 # floci-api config (packages/api)
 FLOCI_ENDPOINT=http://localhost:4566
+VITE_MOCK_MODE=false
 AWS_REGION=us-east-1
 AWS_ACCESS_KEY_ID=test
 AWS_SECRET_ACCESS_KEY=test
